Fix desktop shortcut block.
Was using "subgrid" values for the grid-template property in the shortcuts block; this is not compatible with most browsers, so the layout would sometimes break. I replaced "subgrid" with a more widely supported grid-template definition + inherited the gaps of the desktop-container grid (since the gaps would disappear otherwise). Thanks to Ali (on Discord) for pointing out the broken layout!